created pages directory and moved style.css into this directory
