+++ client/src/App.tsx
@@ -1,16 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import OAuthClickUp from './components/oauth';
+import OAuthClickUp from './pages/oauth';
 import Layout from './components/Layout';
-import Automations from './components/Automations/automations';
+import Automations from './pages/automations';
 import { io, Socket } from 'socket.io-client';
 import { ClientToServerEvents, ServerToClientEvents } from './models/socket';
 import { access } from 'fs';
-import Workspace from './components/workspace';
+import Workspace from './pages/workspace';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { ThemeProvider } from 'react-bootstrap';
-import BasicAuth from './components/basic';
-import TokenAuth from './components/token';
+import BasicAuth from './pages/basic';
+import TokenAuth from './pages/token';
 
 const socket: Socket<ServerToClientEvents, ClientToServerEvents> = io(
   'http://localhost:3002'
@@ -44,7 +44,7 @@ const App: React.FC<{}> = () => {
         <Routes>
           <Route path="/" index element={<Layout />}></Route>
           <Route path="/basic" element={<BasicAuth />}></Route>
-          <Route path="/token" element={<TokenAuth/>}></Route>
+          <Route path="/token" element={<TokenAuth />}></Route>
           <Route path="/oauth" element={<OAuthClickUp />}></Route>
           <Route path="/oauth/success" element={<OAuthClickUp />}></Route>
           <Route
