+++ client/src/pages/token.tsx
