+++ client/src/pages/workspace.tsx
